Refactor out overlay type (#180)
